+++ src/lib.rs
@@ -15,3 +15,6 @@ mod sea_utils;
 
 #[cfg(feature = "with-sea-query")]
 pub use sea_utils::SIden;
+
+#[cfg(feature = "with-rusqlite")]
+pub mod sqlite;